AutoAttack + Focus_Marker

+++ Hero Defense/Assets/Scripts/Stats/Stat.cs
@@ -6,25 +6,25 @@ using UnityEngine;
 public class Stat {
 
     [SerializeField]
-    private int baseValue;
+    private float baseValue;
 
-    public List<int> modifiers = new List<int>();
+    public List<float> modifiers = new List<float>();
 
-    public int GetValue()
+    public float GetValue()
     {
-        int finalValue = baseValue;
+        float finalValue = baseValue;
         modifiers.ForEach(x => finalValue += x);
 
         return finalValue;
     }
 
-    public void AddModifier(int modifier)
+    public void AddModifier(float modifier)
     {
         if (modifier != 0)
             modifiers.Add(modifier);
     }
 
-    public void RemoveModifier (int modifier)
+    public void RemoveModifier (float modifier)
     {
         if (modifier != 0)
             modifiers.Remove(modifier);
